+++ packages/ai-gateway/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @auto-engineer/ai-gateway
 
+## 0.8.4
+
+### Patch Changes
+
+- bump version up
+
+## 0.8.3
+
+## 0.8.2
+
 ## 0.7.0
 
 ### Minor Changes