+++ packages/create-auto-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-auto-app
 
+## 0.8.4
+
+## 0.8.3
+
+## 0.8.2
+
 ## 0.3.0
 
 ### Minor Changes