+++ packages/flow/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @auto-engineer/flow
 
+## 0.8.4
+
+### Patch Changes
+
+- @auto-engineer/file-store@0.8.4
+- @auto-engineer/message-bus@0.8.4
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [3aff24e]
+- Updated dependencies
+  - @auto-engineer/message-bus@0.8.3
+  - @auto-engineer/file-store@0.8.3
+
+## 0.8.2
+
+### Patch Changes
+
+- fixes package.json dependencies
+  - @auto-engineer/file-store@0.8.2
+  - @auto-engineer/message-bus@0.8.2
+
 ## 0.3.0
 
 ### Minor Changes