+++ packages/frontend-checks/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @auto-engineer/frontend-checks
 
+## 0.8.4
+
+### Patch Changes
+
+- @auto-engineer/message-bus@0.8.4
+
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [3aff24e]
+- Updated dependencies
+  - @auto-engineer/message-bus@0.8.3
+
+## 0.8.2
+
+### Patch Changes
+
+- @auto-engineer/message-bus@0.8.2
+
 ## 0.7.0
 
 ### Minor Changes