+++ packages/message-bus/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @auto-engineer/message-bus
 
+## 0.8.4
+
+## 0.8.3
+
+### Patch Changes
+
+- 3aff24e: bump version up
+- version bump
+
+## 0.8.2
+
 ## 0.6.0
 
 ### Minor Changes